requirements.txt
numpy
Pillow
pyyaml
imageio
torch
chromadb
mlflow
